Refactoring for multiple visual editors
