+++ Source/Tooling/vscode/ApplicationEditorProvider.ts
@@ -2,9 +2,9 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 import * as vscode from 'vscode';
-import { getNonce } from './util';
+import { VisualEditor } from './VisualEditor';
 
-export class ApplicationEditorProvider implements vscode.CustomTextEditorProvider {
+export class ApplicationEditorProvider extends VisualEditor {
     private static readonly viewType = 'dolittle.application';
 
     public static register(context: vscode.ExtensionContext): vscode.Disposable {
@@ -13,61 +13,24 @@ export class ApplicationEditorProvider implements vscode.CustomTextEditorProvide
         return providerRegistration;
     }
 
-    constructor(
-        private readonly context: vscode.ExtensionContext
-    ) { }
+    constructor(context: vscode.ExtensionContext) {
+        super(context);
+    }
 
-    async resolveCustomTextEditor(
+    get name(): string {
+        return 'application';
+    }
+
+    async onInitialize(
         document: vscode.TextDocument,
         webviewPanel: vscode.WebviewPanel,
         token: vscode.CancellationToken
     ): Promise<void> {
 
-        const channel = vscode.window.createOutputChannel('Dolittle Application');
-        channel.appendLine('Starting custom text editor');
-        webviewPanel.webview.options = {
-            enableScripts: true
-        };
-        webviewPanel.webview.html = this.getHtmlForWebview(webviewPanel.webview);
 
         webviewPanel.webview.onDidReceiveMessage(e => {
-            channel.appendLine(e.type);
+            this.outputChannel.appendLine(e.type);
             vscode.window.showInformationMessage(e.type);
         });
-
-        function updateWebview() {
-            webviewPanel.webview.postMessage({
-                type: 'update',
-                text: document.getText(),
-            });
-        }
-
-        updateWebview();
-    }
-
-
-    private getHtmlForWebview(webview: vscode.Webview): string {
-        // Local path to script and css for the webview
-        const scriptUri = webview.asWebviewUri(vscode.Uri.joinPath(
-            this.context.extensionUri, 'dist/Visual/index.js'));
-
-        const nonce = getNonce();
-
-        return /* html */`
-            <!DOCTYPE html>
-            <html>
-
-            <head>
-                <meta charset="utf-8" />
-                <title>Dolittle Application</title>
-                <meta name="viewport" content="width=device-width, initial-scale=1">
-            </head>
-
-            <body class="ms-Fabric">
-                <div id="root"></div>
-                <script nonce="${nonce}" src="${scriptUri}"></script>
-            </body>
-
-            </html>`;
     }
 }
